save changes

+++ database/migrations/2023_09_01_154516_create_products_table.php
@@ -27,9 +27,9 @@ class CreateProductsTable extends Migration
             $table->text('tags');
             $table->tinyInteger('status')->default(0);
             $table->tinyInteger('marketable')->default(1);
-            $table->tinyInteger('sold_number')->default(0);
-            $table->tinyInteger('frozen_number')->default(0);
-            $table->tinyInteger('marketable_number')->default(0);
+            $table->bigInteger('sold_number')->default(0);
+            $table->bigInteger('frozen_number')->default(0);
+            $table->bigInteger('marketable_number')->default(0);
             $table->foreignId('brand_id')->constrained('brands')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('category_id')->constrained('product_categories')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamp('published_at');
